Add question to confirm deletion of exp folder
Better when the error is stupid/near the end.

+++ expman.py
@@ -3,7 +3,39 @@ from datetime import datetime
 from subprocess import Popen, PIPE
 import json
 import shutil
+import sys
 import __main__
+
+
+def query_yes_no(question, default="yes"):
+    """Ask a yes/no question via raw_input() and return their answer.
+
+    "question" is a string that is presented to the user.
+    "default" is the presumed answer if the user just hits <Enter>.
+            It must be "yes" (the default), "no" or None (meaning
+            an answer is required of the user).
+
+    The "answer" return value is True for "yes" or False for "no".
+    """
+    valid = {"yes": True, "y": True, "ye": True, "no": False, "n": False}
+    if default is None:
+        prompt = " [y/n] "
+    elif default == "yes":
+        prompt = " [Y/n] "
+    elif default == "no":
+        prompt = " [y/N] "
+    else:
+        raise ValueError("invalid default answer: '%s'" % default)
+
+    while True:
+        sys.stdout.write(question + prompt)
+        choice = input().lower()
+        if default is not None and choice == "":
+            return valid[default]
+        elif choice in valid:
+            return valid[choice]
+        else:
+            sys.stdout.write("Please respond with 'yes' or 'no' " "(or 'y' or 'n').\n")
 
 
 def get_current_hash():
@@ -44,7 +76,9 @@ class ExpLogger:
         # here we catch all exceptions as deadly
         # so we cleanup the whole directory in such case
         if exc_type is not None:
-            shutil.rmtree(self.tmpdir.parent)
+            # ask the user what to do
+            if query_yes_no("Delete experiment folder?"):
+                shutil.rmtree(self.tmpdir.parent)
             return None
         else:
             # zip the contents of the folder
